Move context resource to AbstractParseErrorReporter
The context resource is used by Xtext annexes and should be available in
all parse error reporters.
